+++ tests/test_highlight.py
@@ -2178,7 +2178,8 @@ def test_keyword_collection(invoke, assert_output_regex):
     result = invoke(color_cli1, "command1", "--help", color=True)
     assert result.stdout == (
         "It works!\n"
-        "\x1b[94m\x1b[4mUsage:\x1b[0m \x1b[97m\x1b[1mcolor-cli1 command1\x1b[0m"
+        "\x1b[94m\x1b[4mUsage:\x1b[0m \x1b[97m\x1b[1mcolor-cli1\x1b[0m"
+        " \x1b[36m\x1b[1mcommand1\x1b[0m"
         " \x1b[36m\x1b[2m\x1b[3m[OPTIONS]\x1b[0m \x1b[36m\x1b[3m[MY_ARG]...\x1b[0m\n"
         "\n"
         "  CLI description with extra MY_VAR reference.\n"